+++ shared/components/panel-layout/components/sidebar/components/sidebar-mobile.tsx
@@ -1,0 +1,117 @@
+import { cn } from '@/shared/utils/cn';
+import {
+  X,
+  AlignLeft,
+  ChevronRight,
+  ChevronDown
+} from 'lucide-react';
+import Link from 'next/link';
+import { usePathname } from 'next/navigation';
+import React, { useState } from 'react';
+import Logo from '@/shared/components/icons/Logo';
+import {
+  Drawer,
+  DrawerClose,
+  DrawerContent,
+  DrawerTrigger
+} from '@/shared/components/ui/drawer';
+import { Button } from '@/shared/components/ui/Button';
+import { NavMenuPropsKeys, SIDE_BAR_MENU } from '../../const';
+import { Collapsible, CollapsibleContent, CollapsibleTrigger } from '@/shared/components/ui/collapsible';
+
+const SidebarMobile = () => {
+  const path = usePathname();
+  const [openMenu, setOpenMenu] = useState(Object.keys(SIDE_BAR_MENU)[1]);
+
+  const handleMenuClick = (menu: NavMenuPropsKeys) => {
+    setOpenMenu((prevMenu) => (prevMenu === menu ? '' : menu));
+  };
+
+  return (
+    <div className="lg:hidden">
+      <Drawer direction="left">
+        <DrawerTrigger asChild>
+          <div className="flex items-center w-fit h-20 bg-transparent top-0 absolute z-50 ">
+            <Button className="text-foreground gap-5" variant="link">
+              <AlignLeft className="mb-1.5" />
+            </Button>
+          </div>
+        </DrawerTrigger>
+        <DrawerContent className="flex flex-col h-full w-[300px] mt-24 fixed bottom-0 rounded-none">
+          <div className="p-4 bg-background flex-1 h-full flex flex-col gap-4">
+            <div className="flex gap-2 items-center">
+              <div className="flex items-center w-[40px] h-[40px] min-w-[40px]">
+                <Logo />
+              </div>
+              <h2 className="font-bold text-primary">SWARMS</h2>
+            </div>
+            <DrawerClose className="absolute top-4 right-4">
+              <X />
+            </DrawerClose>
+            {Object.keys(SIDE_BAR_MENU).map((menuKey) => {
+              const menu = menuKey as NavMenuPropsKeys;
+              return (
+                <Collapsible
+                  key={menu}
+                  className="flex-col transition-all duration-300 ease-in-out"
+                  open={openMenu === menu}
+                  onOpenChange={() => handleMenuClick(menu)}
+                >
+                  <CollapsibleTrigger className="justify-between p-2 py-3 my-1 hover:bg-destructive rounded-md hover:text-white outline-none">
+                    <span className="capitalize text-base font-semibold">
+                      {menu}
+                    </span>
+                    {openMenu === menu ? <ChevronDown /> : <ChevronRight />}
+                  </CollapsibleTrigger>
+                  <CollapsibleContent className="flex flex-col gap-1">
+                    {SIDE_BAR_MENU?.[menu]?.map((item, index) => (
+                      <div className="flex flex-col gap-2" key={index}>
+                        <Link
+                          href={item.link}
+                          className={cn(
+                            'group flex items-center justify-start p-2 py-3 my-1 hover:bg-primary hover:text-white rounded-md outline-none',
+                            item.link === path && 'bg-primary text-white'
+                          )}
+                        >
+                          {item.icon && (
+                            <span
+                              className={cn(
+                                'mr-2 text-black dark:text-white group-hover:text-white',
+                                item.link === path && 'text-white'
+                              )}
+                            >
+                              {item.icon}
+                            </span>
+                          )}
+                          <span>{item.title}</span>
+                        </Link>
+                        {item.link === path && item.items?.length && (
+                          <div className="flex flex-col gap-2">
+                            {item.items?.map((subItem) => (
+                              <Link
+                                href={subItem.link}
+                                className={cn(
+                                  'pl-10  py-1 group flex items-center justify-start hover:bg-primary hover:text-white rounded-md outline-none',
+                                  subItem.link === path &&
+                                    'bg-primary dark:text-white'
+                                )}
+                              >
+                                <span>{subItem.title}</span>
+                              </Link>
+                            ))}
+                          </div>
+                        )}
+                      </div>
+                    ))}
+                  </CollapsibleContent>
+                </Collapsible>
+              );
+            })}
+          </div>
+        </DrawerContent>
+      </Drawer>
+    </div>
+  );
+};
+
+export default SidebarMobile;
